# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Atlas Billing API
  version: 0.2.0
  description: Subscription billing — products, prices, coupons, subscriptions, invoices, and credit notes
  contact:
    name: AtlasPayments Team
    url: https://github.com/shivemind/atlaspayments
servers:
  - url: http://localhost:3002
    description: Local billing service. Override via Postman {{baseUrl}} environment variable.
...
paths:
...
  /api/v1/coupons:
    get:
      summary: List coupons
      description: Returns a paginated list of coupons for the authenticated merchant.
      x-platform:
        kind: resource
        entity: Coupon
        affectsLedger: false
      operationId: listCoupons
      tags:
        - Coupons
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Paginated list of coupons.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - pagination
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create coupon
      description: >
        Creates a coupon with either a fixed amount or percentage discount. Exactly one of amount_off or percent_off
        must be provided.
      x-platform:
        kind: resource
        entity: Coupon
        emitsEvents:
          - coupon.created
        affectsLedger: false
      operationId: createCoupon
      tags:
        - Coupons
      security:
        - MerchantKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 200
                amount_off:
                  type: integer
                  minimum: 1
                  description: Fixed discount in smallest currency unit. Mutually exclusive with percent_off.
                percent_off:
                  type: number
                  minimum: 0.01
                  maximum: 100
                  description: Percentage discount. Mutually exclusive with amount_off.
                currency:
                  type: string
                  minLength: 3
                  maxLength: 3
                  description: Required when amount_off is set.
                duration:
                  type: string
                  enum:
                    - once
                    - repeating
                    - forever
                  default: once
                duration_in_months:
                  type: integer
                  minimum: 1
                  description: Required when duration is repeating.
                max_redemptions:
                  type: integer
                  minimum: 1
                redeem_by:
                  type: string
                  format: date-time
                metadata:
                  type: object
                  additionalProperties: true
            examples:
              percent:
                value:
                  name: Summer Sale
                  percent_off: 20
                  duration: once
      responses:
        "201":
          description: Coupon created successfully.
          content:
            application/json:
              schema:
                type: object
                required:
                  - coupon
                properties:
                  coupon:
                    type: object
        "400":
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/coupons/{id}:
    get:
      summary: Get coupon by ID
      description: Returns a single coupon scoped to the authenticated merchant.
      x-platform:
        kind: resource
        entity: Coupon
        affectsLedger: false
      operationId: getCouponById
      tags:
        - Coupons
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Coupon found.
          content:
            application/json:
              schema:
                type: object
                required:
                  - coupon
                properties:
                  coupon:
                    type: object
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Coupon not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      summary: Update coupon
      description: Updates a coupon's name or metadata. Cannot update inactive coupons.
      x-platform:
        kind: transition
        entity: Coupon
        emitsEvents:
          - coupon.updated
        affectsLedger: false
        stateChanges:
          - active -> inactive
          - metadata_updated
      operationId: updateCoupon
      tags:
        - Coupons
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 200
                metadata:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: Coupon updated.
          content:
            application/json:
              schema:
                type: object
                required:
                  - coupon
                properties:
                  coupon:
                    type: object
        "400":
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Coupon not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Coupon is inactive.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Deactivate coupon
      description: Deactivates a coupon (soft delete). Inactive coupons cannot be updated.
      x-platform:
        kind: transition
        entity: Coupon
        stateChanges:
          - active -> inactive
        emitsEvents:
          - coupon.deactivated
        affectsLedger: false
      operationId: deactivateCoupon
      tags:
        - Coupons
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Coupon deactivated.
          content:
            application/json:
              schema:
                type: object
                required:
                  - coupon
                properties:
                  coupon:
                    type: object
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Coupon not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Coupon is already inactive.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/adjustments:
    get:
      summary: List adjustments
      description: Returns a paginated list of adjustments. Optionally filter by type.
      x-platform:
        kind: resource
        entity: Adjustment
        affectsLedger: false
      operationId: listAdjustments
      tags:
        - CreditNotes
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: type
          in: query
          schema:
            type: string
            enum:
              - credit
              - debit
      responses:
        "200":
          description: Paginated list of adjustments.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - pagination
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Adjustment"
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create adjustment
      description: Creates a credit or debit adjustment for the authenticated merchant.
      x-platform:
        kind: resource
        entity: Adjustment
        emitsEvents:
          - adjustment.created
        affectsLedger: false
      operationId: createAdjustment
      tags:
        - CreditNotes
      security:
        - MerchantKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type
                - amount
                - currency
              properties:
                type:
                  type: string
                  enum:
                    - credit
                    - debit
                amount:
                  type: integer
                  minimum: 1
                currency:
                  type: string
                  minLength: 3
                  maxLength: 3
                description:
                  type: string
                  maxLength: 500
                reference:
                  type: string
                  maxLength: 200
                metadata:
                  type: object
                  additionalProperties: true
      responses:
        "201":
          description: Adjustment created.
          content:
            application/json:
              schema:
                type: object
                required:
                  - adjustment
                properties:
                  adjustment:
                    $ref: "#/components/schemas/Adjustment"
        "400":
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    MerchantKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        Bearer API key for merchant-scoped operations. Pass as `Authorization: Bearer {{merchantApiKey}}`.
    PlatformAdminKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        Bearer API key for platform admin operations. Pass as `Authorization: Bearer {{platformAdminKey}}`.
  schemas:
    ErrorResponse:
      description: Standard error response envelope returned for 4xx/5xx responses.
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code.
            message:
              type: string
              description: Human-readable error description.
    PaginationMeta:
      description: Pagination metadata included in list responses.
      type: object
      required:
        - page
        - pageSize
        - total
        - hasMore
      properties:
        page:
          type: integer
          description: Current page number (1-indexed).
        pageSize:
          type: integer
          description: Number of results per page.
        total:
          type: integer
          description: Total number of matching records.
        hasMore:
          type: boolean
          description: Whether additional pages exist.
...
    Adjustment:
      description: A credit or debit adjustment.
      type: object
      properties:
        id:
          type: string
        merchant_id:
          type: string
        type:
          type: string
          enum:
            - credit
            - debit
        amount:
          type: integer
        currency:
          type: string
        description:
          type: string
          nullable: true
        reference:
          type: string
          nullable: true
        metadata:
          type: object
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
  parameters:
...
    PageParam:
      in: query
      name: page
      description: Page number (1-indexed).
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      in: query
      name: pageSize
      description: Number of results per page (max 100).
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    PathId:
      in: path
      name: id
      required: true
      description: Resource identifier.
      schema:
        type: string
...